unify tripuserinput and placeuserinput
- since these are basically the same, they will both inherit from the generic Userinput
